+++ data/julie.txt
@@ -1,0 +1,5 @@
+T | 0 | work
+T | 1 | something
+E | 0 | UTR ResLife julie.task.Event | 12-02-2025 1200 | 12-02-2025 1800
+D | 0 | cs2103t Week 3 iP | 10-02-2025 1159
+T | 0 | something new
