+++ src/main/java/julie/command/Command.java
@@ -1,3 +1,10 @@
+package julie.command;
+
+import julie.Storage;
+import julie.TaskList;
+import julie.UI;
+import julie.exception.WrongFormatException;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, UI ui, Storage storage) throws WrongFormatException;
