+++ src/main/java/julie/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package julie.command;
+
+import julie.Storage;
+import julie.TaskList;
+import julie.UI;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, UI ui, Storage storage) {
