+++ src/main/java/julie/exception/WrongFormatException.java
@@ -1,3 +1,5 @@
+package julie.exception;
+
 public class WrongFormatException extends Exception {
 
     public WrongFormatException(String message) {
