+++ src/main/java/julie/task/Deadline.java
@@ -1,3 +1,7 @@
+package julie.task;
+
+import julie.exception.WrongFormatException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
